rewiring for the registration submit button
